api/requirements.txt
python-dotenv
google-generativeai
fastapi
langchain
langchain-google-genai
redis
langchain-community
langchain-redis
uvicorn
pymongo
pycryptodome